Bug 11090 - Bootstrap theme: fix CSS problem with suggestions toolbar links
There is a problem with the style on the toolbar on the suggestions page
which makes it float incorrectly. This patch corrects the problem.

To test, log in to the OPAC as a user who has suggestions. The
selections toolbar (which shows the "select all," etc. controls) should
look correct.

Signed-off-by: JD@TeTakere <joanned@tetakere.org.nz>
Signed-off-by: Jonathan Druart <jonathan.druart@biblibre.com>
Signed-off-by: Galen Charlton <gmc@esilibrary.com>